ref(ricer::config::locator): Move tests into separate file
